Google_Meet_Transcribe/requirements.txt
google-generativeai
reportlab
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
markdown
python-dotenv
